feat: replace in-memory db with postgres and typeorm in track module

+++ src/track/track.controller.ts
@@ -5,6 +5,7 @@ import {
   Get,
   HttpCode,
   HttpStatus,
+  NotFoundException,
   Param,
   ParseUUIDPipe,
   Post,
@@ -13,8 +14,9 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { TrackService } from './track.service';
-import { Track } from './types/track';
 import { TrackDto } from './dto/track.dto';
+import Track from './entities/track.entity';
+import { ApiNotFoundResponse } from '@nestjs/swagger';
 
 @Controller('track')
 export class TrackController {
@@ -27,14 +29,17 @@ export class TrackController {
   }
 
   @Put(':id')
+  @ApiNotFoundResponse({ description: 'Track was not found.' })
   @UsePipes(new ValidationPipe({ skipNullProperties: true }))
   async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateTrackDto: TrackDto,
   ): Promise<Track> {
     const track = await this.trackService.findOne(id);
-    this.trackService.throwNotFoundException(track, id);
-    return this.trackService.update(id, track, updateTrackDto);
+    if (!track) {
+      throw new NotFoundException('Track was not found.');
+    }
+    return this.trackService.update(id, updateTrackDto);
   }
 
   @Get()
@@ -43,21 +48,27 @@ export class TrackController {
   }
 
   @Get(':id')
+  @ApiNotFoundResponse({ description: 'Track was not found.' })
   async findOne(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
   ): Promise<Track> {
     const track = await this.trackService.findOne(id);
-    this.trackService.throwNotFoundException(track, id);
+    if (!track) {
+      throw new NotFoundException('Track was not found.');
+    }
     return track;
   }
 
   @Delete(':id')
+  @ApiNotFoundResponse({ description: 'Track was not found.' })
   @HttpCode(HttpStatus.NO_CONTENT)
   async remove(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
   ): Promise<void> {
     const track = await this.trackService.findOne(id);
-    this.trackService.throwNotFoundException(track, id);
+    if (!track) {
+      throw new NotFoundException('Track was not found.');
+    }
     await this.trackService.remove(track);
   }
 }
